\name{Associate.Clusters.wcv.2}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
Associate.Clusters.wcv.2(cl, DataA.tbc, DataA.tbp, DataS.tbc)
}
\alias{Associate.Clusters.wcv.2}
\description{
This function is for internal use and is not meant to be called by the user.
}
\arguments{
  \item{cl}{}
  \item{DataA.tbc}{}
  \item{DataA.tbp}{}
  \item{DataS.tbc}{}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\examples{
## Example
}
\keyword{internal}
