\name{cancor.Vrac}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
cancor.Vrac(DataS,DataA)
}
\alias{cancor.Vrac}
\description{
This function is for internal use and is not meant to be called by the user.
It computes a canonical correlation analysis.
}
\arguments{
  \item{DataS}{Dataset 1 (e.g., local-scale at stations).}
  \item{DataA}{Dataset 2 (e.g., large-scale atmospheric data).}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\seealso{\code{\link{CCM}}}
\examples{
## Example
}
\keyword{internal}
