\name{Euclidian.distance}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
Euclidian.distance(X,Y)
}
\alias{Euclidian.distance}
\description{
This function is for internal use and is not meant to be called by the user.
}
\arguments{
  \item{X}{}
  \item{Y}{}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\examples{
## Example
}
\keyword{internal}
