\name{kept.data1D.greater.than.r}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
kept.data1D.greater.than.r(Data,r)
}
\alias{kept.data1D.greater.than.r}
\description{
This function is for internal use and is not meant to be called by the user.
From a vector Data, it returns the vector of values in Data that are < r.
}
\arguments{
  \item{Data}{Vector of numerical values.}
  \item{r}{Value max.}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\examples{
## Example
}
\keyword{internal}
