\name{kept.data2D.from.cluster}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
kept.data2D.from.cluster(cl,c,Nc,Data)
}
\alias{kept.data2D.from.cluster}
\description{
This function is for internal use and is not meant to be called by the user.
Returns the (2d) values contained in one cluster, based on the sequence of clusters.
}
\arguments{
  \item{cl}{Sequence of clusters.}
  \item{c}{Cluster selected.}
  \item{Nc}{Total number of clusters (in the sequence cl).}
  \item{Data}{2D matrix of numerical values.}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\examples{
## Example
}
\keyword{internal}
