\name{Pick.k.random}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
Pick.k.random(k,tab2d)
}
\alias{Pick.k.random}
\description{
This function is for internal use and is not meant to be called by the user.
}
\arguments{
  \item{k}{Number of elements to be picked.}
  \item{tab2d}{2D matrix}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\examples{
## Example
}
\keyword{internal}
