\name{project.on.w.CCA.2}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
project.on.w.CCA.2(DataA.tbp, I, J, B, mA=NaN)
}
\alias{project.on.w.CCA.2}
\description{
This function is for internal use and is not meant to be called by the user.
}
\arguments{
  \item{DataA.tbp}{}
  \item{I}{}
  \item{J}{}
  \item{B}{}
  \item{mA}{}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\examples{
## Example
}
\keyword{internal}
