\name{search.PCAlevel}
\title{This function is for internal use and is not meant to be called by the user.}
\usage{
search.PCAlevel(PCAGL,p)
}
\alias{search.PCAlevel}
\description{
This function is for internal use and is not meant to be called by the user.
}
\arguments{
  \item{PCAGL}{}
  \item{p}{}
}
\author{M. Vrac (mathieu.vrac@lsce.ipsl.fr))}
\examples{
## Example
}
\keyword{internal}
